Подкорректировал расчет крыши, теперь определяется более приближенно к оригиналу. Сделал первую версию расчета отображения источников света. Вынес инициализацию текстур, расчет индекса очереди отрисовки и т.п. в отдельную функцию (для игровых предметов). Прицепил ссылку на тайлдату каждому предмету типа Static.

+++ FrameBuffer.h
@@ -43,7 +43,7 @@ public:
 	bool Ready(int &width, int &height);
 	bool Use();
 	
-	void Draw(float x, float y);
+	void Draw(int x, int y);
 };
 
 extern TFrameBuffer g_LightBuffer;
